refactor: reorganize architecture traits and modules
- Introduced `Arch` trait as a top-level architecture abstraction combining `CpuOps` and `VirtualMemory`.
- Created `CpuOps` trait for low-level CPU operations, facilitating architecture portability.
- Defined `VirtualMemory`, `UserAddressSpace`, and `KernAddressSpace` traits for memory management.
- Updated all architecture implementations (RISC-V, LoongArch) to use the new trait structure.
- Implemented a mock architecture for testing purposes, ensuring compatibility with non-target architectures.
- Refactored imports across the codebase to align with the new architecture module structure.
- Removed the old HAL module structure, consolidating architecture-related traits into the `arch` module.
