+++ os/src/arch/loongarch/cpu_ops.rs
@@ -1,10 +1,10 @@
 //! LoongArch64 CpuOps 实现
 //!
-//! 提供 `LoongArch64` 结构体，实现 `hal::CpuOps` trait。
+//! 提供 `LoongArch64` 结构体，实现 `arch::CpuOps` trait。
 //! 将现有的 CSR 内联汇编操作映射到 trait 方法。
 
 use crate::arch::constant::{CSR_CRMD_IE, SSTATUS_SIE};
-use crate::hal::CpuOps;
+use crate::arch::CpuOps;
 
 /// LoongArch64 架构标记类型
 pub struct LoongArch64;
