+++ os/src/sync/intr_guard.rs
@@ -10,7 +10,7 @@
 //! * `CPU` - 实现 `CpuOps` 的类型，默认使用 `ArchImpl`（编译时选择的架构）
 
 use crate::arch::ArchImpl;
-use crate::hal::CpuOps;
+use crate::arch::CpuOps;
 use core::marker::PhantomData;
 use core::ops::Drop;
