+++ os/src/sync/preempt.rs
@@ -9,7 +9,7 @@ use core::sync::atomic::{AtomicUsize, Ordering};
 
 use crate::arch::ArchImpl;
 use crate::config::MAX_CPU_COUNT;
-use crate::hal::CpuOps;
+use crate::arch::CpuOps;
 
 /// 缓存行对齐的原子计数器
 #[repr(align(64))]
